[SIFISC-1412] Calculer le code OID a afficher sur les déclaration d'impots en fonction de la  Région à laquelle la  commune de gestion du contribuable est rattachée
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@67407 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/core/src/main/java/ch/vd/uniregctb/tiers/CollectiviteAdministrative.java
@@ -35,9 +35,9 @@ public class CollectiviteAdministrative extends Contribuable {
 	 */
 	private Integer numeroCollectiviteAdministrative;
 
-	private Long identifiantDistrictFiscal;
+	private Integer identifiantDistrictFiscal;
 
-	private Long identifiantRegionFiscale;
+	private Integer identifiantRegionFiscale;
 
 
 	/**
@@ -85,20 +85,20 @@ public class CollectiviteAdministrative extends Contribuable {
 	}
 
 	@Column(name = "DISTRICT_FISCAL_ID", nullable = true)
-	public Long getIdentifiantDistrictFiscal() {
+	public Integer getIdentifiantDistrictFiscal() {
 		return identifiantDistrictFiscal;
 	}
 
-	public void setIdentifiantDistrictFiscal(Long identifiantDistrictFiscal) {
+	public void setIdentifiantDistrictFiscal(Integer identifiantDistrictFiscal) {
 		this.identifiantDistrictFiscal = identifiantDistrictFiscal;
 	}
 
 	@Column(name = "REGION_FISCALE_ID", nullable = true)
-	public Long getIdentifiantRegionFiscale() {
+	public Integer getIdentifiantRegionFiscale() {
 		return identifiantRegionFiscale;
 	}
 
-	public void setIdentifiantRegionFiscale(Long identifiantRegionFiscale) {
+	public void setIdentifiantRegionFiscale(Integer identifiantRegionFiscale) {
 		this.identifiantRegionFiscale = identifiantRegionFiscale;
 	}
